[#1319] bug-fix licensing binaries extraction

+++ src/main/java/com/kalessil/phpStorm/phpInspectionsEA/license/LicenseService.java
@@ -24,6 +24,8 @@ import java.util.HashMap;
 
 /* Based on https://wyday.com/limelm/help/using-turboactivate-with-java/ */
 final public class LicenseService {
+    final private String taVersioning = "201905151055";
+
     private int trialDaysRemaining = 0;
     private int licenseDaysRemaining = 0;
 
@@ -74,24 +76,28 @@ final public class LicenseService {
             throw new RuntimeException("Licensing related resources are missing.");
         }
 
-        final Path ideTempFolder     = Paths.get(PathManager.getTempPath());
-        final Path tempFolder        = Files.createTempDirectory(ideTempFolder, "ea-ultimate-").toAbsolutePath();
-        final String[] sourceDetails = binaries.toURI().toString().split("!");
-        final FileSystem pluginJarFs = FileSystems.newFileSystem(URI.create(sourceDetails[0]), new HashMap<>());
-        Files.walk(pluginJarFs.getPath(sourceDetails[1])).forEach(sourceFile -> {
-            try {
-                Files.copy(
-                    sourceFile,
-                    tempFolder.resolve(tempFolder.toString() + File.separator + sourceFile.toString()),
-                    StandardCopyOption.COPY_ATTRIBUTES
-                );
-            } catch (Throwable copyFailure) {
-                throw new RuntimeException(copyFailure);
-            }
-        });
-        pluginJarFs.close();
-
-        this.client = new TurboActivate("2d65930359df9afb6f9a54.36732074", tempFolder.toString() + "/TurboActivate/");
+        final Path location = (new File(Paths.get(PathManager.getTempPath()).toFile(), "ea-ultimate-" + taVersioning)).toPath();
+        final Path absolute = location.toAbsolutePath();
+        final String path   = absolute.toString();
+        if (!Files.exists(location)) {
+            /* create location, copy TA resources */
+            Files.createDirectory(location);
+            final String[] sourceDetails = binaries.toURI().toString().split("!");
+            final FileSystem pluginJarFs = FileSystems.newFileSystem(URI.create(sourceDetails[0]), new HashMap<>());
+            Files.walk(pluginJarFs.getPath(sourceDetails[1])).forEach(sourceFile -> {
+                try {
+                    Files.copy(
+                            sourceFile,
+                            absolute.resolve(path + File.separator + sourceFile.toString()),
+                            StandardCopyOption.COPY_ATTRIBUTES
+                    );
+                } catch (Throwable copyFailure) {
+                    throw new RuntimeException(copyFailure);
+                }
+            });
+            pluginJarFs.close();
+        }
+        this.client = new TurboActivate("2d65930359df9afb6f9a54.36732074", path + "/TurboActivate/");
     }
 
     boolean isClientInitialized() {
